Move getEventsFromEditorStateTransfer to a separate file
